server/python/requirements.txt
# SatyaAI Python Backend - Unified Requirements
# Production-ready dependencies for deepfake detection

# ============================================
# Core Framework (FastAPI)
# ============================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.21
pydantic>=2.5.3
pydantic-settings>=2.1.0

# ============================================
# Database & Caching
# ============================================
sqlalchemy>=2.0.25
redis>=5.0.1

# ============================================
# Authentication & Security
# ============================================
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# ============================================
# Deep Learning & AI
# ============================================
torch>=2.0.1
torchvision>=0.15.2
torchaudio>=2.0.2
tensorflow>=2.20.0
keras>=3.0.0

# ============================================
# Computer Vision
# ============================================
opencv-python>=4.9.0.80
opencv-contrib-python>=4.9.0.80
Pillow>=10.2.0
scikit-image>=0.22.0
mediapipe>=0.10.9

# ============================================
# Audio Processing
# ============================================
librosa>=0.10.1
soundfile>=0.12.1
pydub>=0.25.1
scipy>=1.11.4
numpy>=1.24.3

# ============================================
# Video Processing
# ============================================
moviepy>=1.0.3
imageio>=2.33.1
imageio-ffmpeg>=0.4.9

# ============================================
# ML Model Management
# ============================================
transformers>=4.36.2
timm>=0.9.12
efficientnet-pytorch>=0.7.1
facenet-pytorch>=2.5.3
speechbrain>=0.5.16

# ============================================
# Data Processing
# ============================================
pandas>=2.1.4
scikit-learn>=1.3.2
matplotlib>=3.8.2
seaborn>=0.13.1

# ============================================
# API & Networking
# ============================================
requests>=2.32.5
urllib3>=2.1.0
aiohttp>=3.9.1
httpx>=0.26.0  # Required for TestClient

# ============================================
# Utilities
# ============================================
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.1
colorama>=0.4.6

# ============================================
# Testing (Development)
# ============================================
pytest>=8.2.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1

# ============================================
# Code Quality (Development)
# ============================================
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pylint==3.0.3

# ============================================
# Performance Monitoring
# ============================================
psutil==5.9.6
memory-profiler==0.61.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
